Track attachment changes centrally
