--- activestorage/lib/active_storage/attached/macros.rb
@@ -1,156 +1,0 @@
-# frozen_string_literal: true
-
-module ActiveStorage
-  # Provides the class-level DSL for declaring that an Active Record model has attached blobs.
-  module Attached::Macros
-    # Specifies the relation between a single attachment and the model.
-    #
-    #   class User < ActiveRecord::Base
-    #     has_one_attached :avatar
-    #   end
-    #
-    # There is no column defined on the model side, Active Storage takes
-    # care of the mapping between your records and the attachment.
-    #
-    # To avoid N+1 queries, you can include the attached blobs in your query like so:
-    #
-    #   User.with_attached_avatar
-    #
-    # Under the covers, this relationship is implemented as a +has_one+ association to a
-    # ActiveStorage::Attachment record and a +has_one-through+ association to a
-    # ActiveStorage::Blob record. These associations are available as +avatar_attachment+
-    # and +avatar_blob+. But you shouldn't need to work with these associations directly in
-    # most circumstances.
-    #
-    # The system has been designed to having you go through the ActiveStorage::Attached::One
-    # proxy that provides the dynamic proxy to the associations and factory methods, like +attach+.
-    #
-    # If the +:dependent+ option isn't set, the attachment will be purged
-    # (i.e. destroyed) whenever the record is destroyed.
-    def has_one_attached(name, dependent: :purge_later)
-      generated_association_methods.class_eval <<-CODE, __FILE__, __LINE__ + 1
-        def #{name}
-          @active_storage_attached_#{name} ||= ActiveStorage::Attached::One.new("#{name}", self, dependent: #{dependent == :purge_later ? ":purge_later" : "false"})
-        end
-
-        def #{name}=(attachable)
-          @active_storage_attached_#{name}_change =
-            if attachable.nil?
-              ActiveStorage::Attached::Changes::DeleteOne.new("#{name}", self)
-            else
-              ActiveStorage::Attached::Changes::CreateOne.new("#{name}", self, attachable)
-            end
-        end
-
-        def #{name}_change
-          @active_storage_attached_#{name}_change
-        end
-
-        def clear_#{name}_change
-          @active_storage_attached_#{name}_change = nil
-        end
-      CODE
-
-      has_one :"#{name}_attachment", -> { where(name: name) }, class_name: "ActiveStorage::Attachment", as: :record, inverse_of: :record, dependent: :destroy
-      has_one :"#{name}_blob", through: :"#{name}_attachment", class_name: "ActiveStorage::Blob", source: :blob
-
-      scope :"with_attached_#{name}", -> { includes("#{name}_attachment": :blob) }
-
-      after_save { public_send("#{name}_change")&.save }
-
-      after_commit(on: [:create, :update]) do
-        begin
-          public_send("#{name}_change").try(:upload)
-        ensure
-          public_send("clear_#{name}_change")
-        end
-      end
-
-      ActiveRecord::Reflection.add_attachment_reflection(
-        self,
-        name,
-        ActiveRecord::Reflection.create(:has_one_attached, name, nil, { dependent: dependent }, self)
-      )
-    end
-
-    # Specifies the relation between multiple attachments and the model.
-    #
-    #   class Gallery < ActiveRecord::Base
-    #     has_many_attached :photos
-    #   end
-    #
-    # There are no columns defined on the model side, Active Storage takes
-    # care of the mapping between your records and the attachments.
-    #
-    # To avoid N+1 queries, you can include the attached blobs in your query like so:
-    #
-    #   Gallery.where(user: Current.user).with_attached_photos
-    #
-    # Under the covers, this relationship is implemented as a +has_many+ association to a
-    # ActiveStorage::Attachment record and a +has_many-through+ association to a
-    # ActiveStorage::Blob record. These associations are available as +photos_attachments+
-    # and +photos_blobs+. But you shouldn't need to work with these associations directly in
-    # most circumstances.
-    #
-    # The system has been designed to having you go through the ActiveStorage::Attached::Many
-    # proxy that provides the dynamic proxy to the associations and factory methods, like +#attach+.
-    #
-    # If the +:dependent+ option isn't set, all the attachments will be purged
-    # (i.e. destroyed) whenever the record is destroyed.
-    def has_many_attached(name, dependent: :purge_later)
-      generated_association_methods.class_eval <<-CODE, __FILE__, __LINE__ + 1
-        def #{name}
-          @active_storage_attached_#{name} ||= ActiveStorage::Attached::Many.new("#{name}", self, dependent: #{dependent == :purge_later ? ":purge_later" : "false"})
-        end
-
-        def #{name}=(attachables)
-          @active_storage_attached_#{name}_change =
-            if attachables.nil? || Array(attachables).none?
-              ActiveStorage::Attached::Changes::DeleteMany.new("#{name}", self)
-            else
-              ActiveStorage::Attached::Changes::CreateMany.new("#{name}", self, attachables)
-            end
-        end
-
-        def #{name}_change
-          @active_storage_attached_#{name}_change
-        end
-
-        def clear_#{name}_change
-          @active_storage_attached_#{name}_change = nil
-        end
-      CODE
-
-      has_many :"#{name}_attachments", -> { where(name: name) }, as: :record, class_name: "ActiveStorage::Attachment", inverse_of: :record, dependent: :destroy do
-        def purge
-          each(&:purge)
-          reset
-        end
-
-        def purge_later
-          each(&:purge_later)
-          reset
-        end
-      end
-      has_many :"#{name}_blobs", through: :"#{name}_attachments", class_name: "ActiveStorage::Blob", source: :blob
-
-      scope :"with_attached_#{name}", -> { includes("#{name}_attachments": :blob) }
-
-      after_save { public_send("#{name}_change")&.save }
-
-      after_commit(on: [:create, :update]) do
-        begin
-          public_send("#{name}_change").try(:upload)
-        ensure
-          public_send("clear_#{name}_change")
-        end
-      end
-
-      ActiveRecord::Reflection.add_attachment_reflection(
-        self,
-        name,
-        ActiveRecord::Reflection.create(:has_many_attached, name, nil, { dependent: dependent }, self)
-      )
-    end
-  end
-end
